inference/pyproject.toml
[project]
name = "image-captioning-inference"
version = "1.0.0"
description = "Production inference service for image captioning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "pillow>=12.0.0",
    "numpy>=1.24.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]
